fix(root): script clean-up
