+++ package.json
@@ -66,9 +66,9 @@
     "access": "public"
   },
   "scripts": {
-    "build": "nub scripts/build-dist.ts",
-    "build:bin": "nub scripts/build-bin.ts",
-    "build-dist": "nub scripts/build-dist.ts",
+    "build": "nub scripts/build/index.ts dist",
+    "build:bin": "nub scripts/build/index.ts bin",
+    "build-dist": "nub scripts/build/index.ts dist",
     "clean-dist": "rm -rf dist",
     "clean": "rm -rf dist && cd page/app && rm -rf dist",
     "cmd:node:test-fail-run": "node dist/cli.js --config './tests/unit/fixtures/test-fail-package.json' --files 'test-fail-package.json' --rootDir './tests/unit/fixtures/' --debug",
@@ -95,7 +95,7 @@
     "lint:fix": "NODE_OPTIONS= oxlint src --fix",
     "hooks:setup": "sh scripts/setup.sh",
     "prepare": "sh scripts/setup.sh",
-    "prepublishOnly": "rm -rf dist && nub scripts/build-dist.ts",
+    "prepublishOnly": "rm -rf dist && nub scripts/build/index.ts dist",
     "preview:table": "nub scripts/preview-version-table.ts",
     "preview:styleguide": "nub src/cli/index.ts --styleguide",
     "preview": "cd page/app && nub run preview",
@@ -115,7 +115,7 @@
     "release:tag": "nub scripts/release/index.ts tag",
     "release:tag:dry": "nub scripts/release/index.ts tag --dry-run",
     "release:dry": "nub scripts/release/index.ts --dry-run",
-    "setup": "nub install && sh scripts/setup.sh && nub scripts/build-dist.ts",
+    "setup": "nub install && sh scripts/setup.sh && nub scripts/build/index.ts dist",
     "test": "nub scripts/run/index.ts",
     "test:e2e:binary": "./tests/e2e/test-binary.sh",
     "update": "nub update",
