+++ scripts/release/types.ts
@@ -105,3 +105,55 @@ export interface ReleaseContext {
   runner: ReleaseRunner;
   schemaMetadataWriter: SchemaMetadataWriter;
 }
+
+export type Fetch = typeof fetch;
+
+export interface FormulaSource {
+  digest: string;
+  url: string;
+}
+
+export interface FormulaInput extends FormulaSource {
+  version: string;
+}
+
+export interface FormulaOptions {
+  outputPath: string;
+  version: string;
+}
+
+export interface PublishedFormulaOptions extends FormulaOptions {
+  fetchImpl?: Fetch;
+}
+
+export interface LocalFormulaOptions extends FormulaOptions {
+  tarballPath: string;
+}
+
+export interface BrewCliOptions {
+  argv?: string[];
+  env?: Record<string, string | undefined>;
+  fetchImpl?: Fetch;
+}
+
+export interface HomebrewReleaseStateOptions {
+  arch?: string;
+  env: Record<string, string | undefined>;
+  fetchImpl?: Fetch;
+}
+
+export interface HomebrewReleaseState {
+  reason?: string;
+  skip: boolean;
+}
+
+export interface HomebrewTapUpdateOptions {
+  env: Record<string, string | undefined>;
+  fetchImpl?: Fetch;
+}
+
+export interface HomebrewTapUpdateResult {
+  branch?: string;
+  changed: boolean;
+  pullRequestUrl?: string;
+}
